Move blockchain interfacing code to ethereum.interface
