+++ mythril/ethereum/interface/rpc/exceptions.py
